+++ react/typings/global.d.ts
@@ -1,1 +1,2 @@
 declare module 'vtex.styleguide'
+declare module '*.png'